pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moatless-mcp-server"
version = "0.1.0"
description = "MCP Server for Moatless Tools - Advanced code analysis and editing capabilities"
readme = "README.md"
authors = [
    {name = "MCP Integration", email = "integration@example.com"},
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "gitpython>=3.1.0",
    "anyio>=4.0.0",
    "requests>=2.28.0",
    # Vector search and embeddings
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-vector-stores-faiss>=0.1.0",
    # Text processing and parsing
    "tiktoken>=0.5.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-languages>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
moatless-mcp-server = "moatless_mcp.server:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true